frontend/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
mongoengine==0.29.1
PyJWT==2.10.1
Werkzeug==3.1.3
